+++ src/entities/shelter/model/shelter.ts
@@ -1,7 +1,4 @@
-// TODO-cch: type 한곳으로 옮기기
-
-export type ShelterType = 'animalHospital' | 'shelter' | 'rescueGroup';
-export type AnimalType = 'cat' | 'dog' | 'other';
+import { Pagination } from '../../common/model';
 
 export interface Shelter {
   name: string;
@@ -12,30 +9,6 @@ export interface Shelter {
   operatingHours: string;
 }
 
-export interface Pagination {
-  totalCount: number;
-  currentPage: number;
-  pageSize: number;
-  totalPages: number;
-}
-
-export interface Shelters {
+export interface Shelters extends Pagination {
   shelters: Shelter[];
-  pagination: Pagination;
 }
-
-export const createShelter = (
-  name: string,
-  type: string,
-  animalType: string,
-  address: string,
-  phoneNumber: string,
-  operatingHours: string,
-): Shelter => ({
-  name,
-  type,
-  animalType,
-  address,
-  phoneNumber,
-  operatingHours,
-});
